backend/requirements.txt
fastapi
uvicorn
google-genai
dotenv
python-multipart
